fix: change return type of SizeofNextLevel method from size_t to uint64_t

+++ src/vm/cache/cache.cpp
@@ -112,7 +112,7 @@ void Cache::WriteByteToNextLevel(uint64_t address, uint8_t value)
     }
 }
 
-size_t Cache::SizeofNextLevel() const
+uint64_t Cache::SizeofNextLevel() const
 {
     if (next_level_cache_)
     {
